Added Tabs, Menu for Lamba and magnetic pendulum. Added stats to rendered image. Some minor refactoring (imagecontrolwidget now creates the generator).
Known bugs:

* On some occasions there is a deadlock that the last thread does not terminate
* The last thread takes very long time (0,5 seconds) to finish

+++ graphics/colorpalette.cpp
@@ -3,26 +3,26 @@
 #include <cmath>
 #include <map>
 
-void ColorPalette::color(float x, float& r, float& g, float& b, float& a) const {
+void ColorPalette::color(double x, double& r, double& g, double& b, double& a) const {
 	r = r_(x);
 	g = g_(x);
 	b = b_(x);
 	a = a_(x);
 }
 
-Spline<float>& ColorPalette::r() {
+Spline<double>& ColorPalette::r() {
 	return r_;
 }
 
-Spline<float>& ColorPalette::g() {
+Spline<double>& ColorPalette::g() {
 	return g_;
 }
 
-Spline<float>& ColorPalette::b() {
+Spline<double>& ColorPalette::b() {
 	return b_;
 }
 
-Spline<float>& ColorPalette::a() {
+Spline<double>& ColorPalette::a() {
 	return a_;
 }
